bau - Read metadata location from json file rather than yml
- The application uses the json file but the ruby script currently reads
from yml. To keep the files in sync and to make it clearer what file to update,
make the ruby script read from the json file as well.
- The environment variable is already set in jenkins job builder, which is
how it finds the correct file for each environment
- All metadata we use at the moment is public and therefore there is no
need to supply the tls cert. Remove reading it from the ruby script.
- Remove the now redundant yml files containing the country metadata